livekit-agent/requirements.txt
livekit-agents[deepgram,silero]>=1.2.0,<1.3.0
python-dotenv>=1.0.0
